requirements.txt
ffmpeg
sounddevice
emoji
openai-whisper
xinference[all]